Fixes to chords and rests implementation
Chords and rests are now handles correctly - roots of chords are used. Simultaneous notes (not defined as chords) are treated as before, with a note chosen randomly. Rests are now included in the duration of notes when they are stretched to fill the time. Library seems to have  a bug here, but I've made a work-around.